Refactor(agent-web-ui): extract startPromptStream helper out of ChatPanel
Pulls the per-instance prompt + chat-mirror plumbing
(user bubble → agent bubble(s) → tool bubbles → query bubbles →
cost annotation → onError surface) out of ChatPanel.onSubmit into a
new composables/promptStreaming.ts. ChatPanel.onSubmit shrinks to a
file-encode + helper call.

Done now because the upcoming multi-select fan-out feature needs the
same chat-mirroring behaviour for each fanned-out agent. The helper
is the only sane way to share it without ~80 lines of copy-paste in
the new MultiSelectBar.

No behaviour change for ChatPanel — every callback, every store
write, every per-bubble tweak is preserved verbatim. The cc-session
cost-bump on the session card still fires.
